Fixed makefile error when volare is not in the path
Before, the makefile was installing `volare` via python, but it
was invoking it as a binary (i.e. assuming it was in the path).
This was causing issues when `volare` was not in the path, or
when several python interpreters were present on the system.
Using `python3 -m volare ...` ensures `volare` is run from the same
python interpreter that installed it, and does not require that
`volare` be in the path.